work with deal form; add userform, listdeal form, user-room form, fixed bugs, add commands

+++ src/main/java/by/epam/task/service/impl/DealServiceImpl.java
@@ -37,14 +37,14 @@ public class DealServiceImpl implements DealService {
 	}
 	
 	@Override
-	public Deal findOneByNickname(String nickname) throws ServiceException {
-		Deal deal = null;
+	public List<Deal> findAllByNickname(String nickname) throws ServiceException {
+		List<Deal> list = null;
 		try {
-			deal = dealDAO.findOneByNickname(nickname);
+			list = dealDAO.findAllByNickname(nickname);
 		} catch (DAOException e) {
 			throw new ServiceException("Error execution findOneByNickname method", e);
 		}
-		return deal;
+		return list;
 	}
 
 	@Override
@@ -61,6 +61,10 @@ public class DealServiceImpl implements DealService {
 	@Override
 	public void delete(int id) throws ServiceException {
 		try {
+			Deal deal = dealDAO.findOne(id);
+			if (deal == null || deal.getState() != DealState.CREATED) {
+				throw new ServiceException("Incorrect state of deal");
+			}
 			int count = dealDAO.delete(id);
 			if (count == 0) {
 				throw new ServiceException("Delete error");
@@ -69,9 +73,9 @@ public class DealServiceImpl implements DealService {
 			throw new ServiceException("Error execution delete method", e);
 		}
 	}
-
+	
 	@Override
-	public int modify(int id, String nickname, int carId, Date dateFrom, Date dateTo, String comment) throws ServiceException {
+	public int modify(int id, String nickname, int carId, Date dateFrom, Date dateTo, String comment, String passportNumber) throws ServiceException {
 		try {
 			CarDAO carDAO = DAOFactory.getInstance().getCarDAO();
 			UserDAO userDAO = DAOFactory.getInstance().getUserDAO();			
@@ -81,18 +85,17 @@ public class DealServiceImpl implements DealService {
 			if (car == null || user == null) {
 				throw new ServiceException("Incorrect identificators");
 			}
-			// Check user access
+			// Check state
 			if (id != 0) {
 				Deal deal = dealDAO.findOne(id);
-				String accessNickname = deal.getUser().getNickname();
-				if (!accessNickname.equals(nickname) || deal.getState() != DealState.CREATED) {
+				if (deal.getState() != DealState.CREATED) {
 					throw new ServiceException("The user does not have rights to edit deal");
 				}
 			}
 			// Creating deal
 			float cost =  car.getPrise() * (dateTo.getTime() - dateFrom.getTime()) / 3600000;
 			Deal deal = new Deal(id, cost, dateFrom, dateTo, comment, null, 
-					new User(user.getId()), new Car(carId), null, DealState.CREATED);
+					new User(user.getId()), new Car(carId), null, DealState.CREATED, passportNumber);
 			if (id == 0) {
 				id = dealDAO.insert(deal);
 			} else {
@@ -119,6 +122,21 @@ public class DealServiceImpl implements DealService {
 			dealDAO.update(deal);
 		} catch (DAOException e) {
 			throw new ServiceException("Error execution confirmDeal method", e);
+		}
+	}
+	
+	@Override
+	public void complete(int id) throws ServiceException {
+		try {
+			Deal deal = dealDAO.findOne(id);
+			// Data validation
+			if (deal == null || deal.getState() != DealState.PAID) {
+				throw new ServiceException("Incorrect state of deal");
+			}
+			deal.setState(DealState.COMPLETED);
+			dealDAO.update(deal);
+		} catch (DAOException e) {
+			throw new ServiceException("Error execution completeDeal method", e);
 		}
 	}
 	
@@ -158,22 +176,36 @@ public class DealServiceImpl implements DealService {
 	}
 	
 	@Override
-	public int addDamage(int id, float bill, String description) throws ServiceException {
+	public void addDamage(int id, float cost, String description) throws ServiceException {
 		DamageDAO damageDAO = DAOFactory.getInstance().getDamageDAO();
-		Damage damage = new Damage(id, bill, description);
+		Damage damage = new Damage(id, cost, description);
 		try {
-			if (id == 0) {
-				id = damageDAO.insert(damage);
-			} else {
-				damageDAO.update(damage);
+			Deal deal = dealDAO.findOne(id);
+			// Data validation
+			if (deal == null || deal.getState() != DealState.PAID) {
+				throw new ServiceException("Incorrect state of deal");
 			}
-			if (id == 0) {
-				throw new ServiceException("Save or update error");
-			}
+			deal.setState(DealState.DAMAGED);
+			dealDAO.update(deal);
+			damageDAO.insert(damage);
 		} catch (DAOException e) {
 			throw new ServiceException("Error executing the addDamage method", e);
 		}
-		return id;
+	}
+	
+	@Override
+	public boolean checkUser(String nickname, int dealId) throws ServiceException {
+		Deal deal = null;
+		try {
+			deal = dealDAO.findOne(dealId);
+		} catch (DAOException e) {
+			throw new ServiceException("Error execution the checkUser command", e);
+		}
+		// Data validation
+		if (deal == null || !deal.getUser().getNickname().equals(nickname)) {
+			return false;
+		}
+		return true;
 	}
 
 }
